fix metrics

+++ app/include/Metrics.h
@@ -14,16 +14,16 @@ struct Metrics
     double frameTimeWindow = 5.0f;
 
     std::deque<double> frameTimes;
-    std::unordered_map<std::string, int64_t> timer;
+    std::unordered_map<std::string, double> timer;
 };
 
 #ifdef NOPROFILE
     #define TIME(metrics, name, func) func;
 #else
     #define TIME(metrics, name, func) { \
-        int64_t start = core::Application::get().getTime(); \
+        double start = core::Application::get().getTime(); \
         func; \
-        int64_t end = core::Application::get().getTime(); \
+        double end = core::Application::get().getTime(); \
         metrics.timer[name] = end - start; \
     }
 #endif
